[update] ディレクトリの場所を変更
