+++ bot/.env.sample
