+++ bot/bots/bot.go
