+++ bot/bots/nleaveban/nleaveban.go
