+++ bot/lib/discord/discord.go
