+++ bot/lib/dotenv/dotenv.go
