+++ docker-compose.yml
@@ -9,18 +9,21 @@ services:
     working_dir: /src
   build_arm:
     <<: *go
+    user: "${UID}:${GID}"
     entrypoint: sh ./build/make_arm.sh
     volumes:
       - "./:/src"
     working_dir: /src
   build_arm64:
     <<: *go
+    user: "${UID}:${GID}"
     entrypoint: sh ./build/make_arm64.sh
     volumes:
       - "./:/src"
     working_dir: /src
   build_amd64:
     <<: *go
+    user: "${UID}:${GID}"
     entrypoint: sh ./build/make_amd64.sh
     volumes:
       - "./:/src"
